Maxime

+++ ineedpower/src/componants/Login.js
@@ -3,27 +3,56 @@ import { Redirect } from "react-router-dom";
 import NewAccount from "./NewAccount";
 import GoogleLogin from "./GoogleLogin";
 
-
 class Signup extends React.Component {
     constructor(props) {
         super(props)
         this.state = {
             email: "",
             password: "",
+            users: {},
+            fetched:false,
+            exist:false,
             Redirect:false
         }
         this.handleSubmit = this.handleSubmit.bind(this);
 
     }
 
+    componentDidMount() {
+        fetch(`http://localhost:5000/users`)
+            .then(res => res.json())
+            .then(res => this.setState({ users: res, fetched: true }));
+    }
+
+    update(e) {
+        this.setState({
+            email: this.refs.email.value,
+            password: this.refs.password.value,
+        });
+    }
+
     handleSubmit(e) {
         e.preventDefault();
         //checking if the data is vallid
         //if the data is valid create a session 
-        sessionStorage.setItem("userData", "LoggedIn");
-        this.setState({
-            Redirect:true
-        })
+
+        const email=this.state.email;
+        const password=this.state.password;
+
+        for(let index=0;index<this.state.users.length;index++){
+            if(email === this.state.users[index].email && password === this.state.password) {
+                this.setState({
+                    exist:true
+                })
+            }
+        }
+        
+        if(this.state.exist) {
+            sessionStorage.setItem("userData", "LoggedIn");
+            this.setState({
+                Redirect:true
+            })
+        }
     }
 
     render() {
@@ -33,8 +62,8 @@ class Signup extends React.Component {
         return (
             <div>
                 <form onSubmit={this.handleSubmit}>
-                    <input type="email" placeholder="email" />
-                    <input type="password" placeholder="password"/>
+                    <input type="email" placeholder="email" ref="email" onChange={this.update.bind(this)} />
+                    <input type="password" placeholder="password" ref="password" onChange={this.update.bind(this)}/>
                     <button type="submit">Log in</button>
                 </form>
             </div>
@@ -50,11 +79,18 @@ class Login extends React.Component {
             version: "default"  //shows the 3 different login options (signUp,NewAccount button and google button) , google is not a priority
         }
         this.handleNewAccount = this.handleNewAccount.bind(this);
+        this.changeVersion = this.changeVersion.bind(this);
     }
 
     handleNewAccount() {
         this.setState({
             version: "newAccount"
+        });
+    }
+
+    changeVersion() {
+        this.setState({
+            version:"default"
         });
     }
 
@@ -76,7 +112,7 @@ class Login extends React.Component {
         }
         if (this.state.version === "newAccount") {
             return (
-                <NewAccount></NewAccount>
+                <NewAccount changeVersion={this.changeVersion}></NewAccount>
             );
         }
         if (this.state.version === "google") {
